mqsv: support for SAF SaTimeT timeout variable in MDS transport [#1658] Issue : In current implementation their is a mismatch of variable sizes between SAF API accepting SaTimeT variable type for timeout ( 64 bit) and MDS transport implementation accepting (32 bit) timeout variable , the MDS transport should adopt the size of the timeout parameter from 32 bits to 64 bits.
Fix :
Now MDS provide interface to accept SaTimeT type timeout variable from SAF APIS

+++ osaf/services/saf/mqsv/mqd/mqd_tmr.c
@@ -82,13 +82,13 @@ void mqd_timer_expiry(NCSCONTEXT uarg)
  *****************************************************************************/
 uint32_t mqd_tmr_start(MQD_TMR *tmr, SaTimeT duration)
 {
-	TRACE_1("The timer started with the duration as %u", (uint32_t)duration);
+	TRACE_1("The timer started with the duration as %lld", duration);
 	if (tmr->tmr_id == TMR_T_NULL) {
 		m_NCS_TMR_CREATE(tmr->tmr_id, duration, mqd_timer_expiry, (void *)tmr);
 	}
 
 	if (tmr->is_active == false) {
-		m_NCS_TMR_START(tmr->tmr_id, (uint32_t)duration, mqd_timer_expiry, (void *)tmr);
+		m_NCS_TMR_START(tmr->tmr_id, duration, mqd_timer_expiry, (void *)tmr);
 		tmr->is_active = true;
 	}
